rename a file

+++ remove_non_results_files.py
@@ -2,12 +2,14 @@ import os
 import glob
 
 
-def remove_previous_files():
+def remove_non_results_files():
     # with open('results.csv') as f:
     #     # count the total number of lines in the file
     #     total_lines = sum(1 for line in f)
     #     print('Total number of lines in the results.csv file: ', total_lines)
     for file in glob.glob('*.csv'):
+        if file == 'results.csv':
+            continue
         os.remove(file)
     for file in glob.glob('cached_mutations_*'):
         os.remove(file)
@@ -16,4 +18,4 @@ def remove_previous_files():
 
 
 if __name__ == '__main__':
-    remove_previous_files()
+    remove_non_results_files()
